some fixes, some logic updates, implemented jwt authentication

+++ src/main/java/com/svartvalp/GameMate/Controllers/UserController.java
@@ -1,19 +1,19 @@
 package com.svartvalp.GameMate.Controllers;
 
-import com.svartvalp.GameMate.Exceptions.ValidationException;
+import com.svartvalp.GameMate.JWT.JWTUtils;
 import com.svartvalp.GameMate.Models.UpdatePasswordMessage;
 import com.svartvalp.GameMate.Models.User;
 import com.svartvalp.GameMate.Services.IUserService;
 import com.svartvalp.GameMate.Validation.Validator;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.security.Principal;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 
 @RestController
@@ -22,14 +22,24 @@ public class UserController {
 
     IUserService userService;
     Validator<User> userValidator;
+    JWTUtils jwtUtils;
 
     @Autowired
-    public UserController(IUserService userService, Validator<User> userValidator) {
+    public void setUserService(IUserService userService) {
         this.userService = userService;
+    }
+
+    @Autowired
+    public void setUserValidator(Validator<User> userValidator) {
         this.userValidator = userValidator;
     }
 
-    @PostMapping(value = "", consumes = MediaType.APPLICATION_JSON_VALUE)
+    @Autowired
+    public void setJwtUtils(JWTUtils jwtUtils) {
+        this.jwtUtils = jwtUtils;
+    }
+
+    @PostMapping(value = "/register", consumes = MediaType.APPLICATION_JSON_VALUE)
     public Mono<User> createUser(@RequestBody User user) {
         userValidator.validate(user, List.of("email", "password", "nickname"));
         return userService.createUser(user);
@@ -41,15 +51,26 @@ public class UserController {
     }
 
     @DeleteMapping(value = "")
-    public void deleteUserByNickname(@RequestBody User user) {
+    public void deleteUserByNickname(@RequestBody User user, Principal principal) {
         userValidator.validate(user, List.of("email", "password", "nickname"));
-        userService.deleteUser(user);
+        if(principal.getName().equals(user.getNickname())) {
+            userService.deleteUser(user);
+        }
     }
 
     @PutMapping(value = "/verify")
-    public Mono<ResponseEntity<?>> verifyUser(@RequestBody User user) {
+    public Mono<Map<String, String>> verifyUser(@RequestBody User user) {
         userValidator.validate(user, List.of("email", "password"));
-        return userService.verifyUser(user).map(bool -> bool? ResponseEntity.ok("ok") : ResponseEntity.status(HttpStatus.FORBIDDEN).body("not ok"));
+        return userService.verifyUser(user).flatMap(verifiedUser -> {
+            Map<String, String> body = new HashMap<>();
+            body.put("token", jwtUtils.createToken(user.getNickname()));
+            return Mono.just(body);
+        });
+    }
+
+    @GetMapping(value = "/{nickname}")
+    public Mono<User> find(@PathVariable("nickname") String nickname) {
+        return userService.findUserByNickName(nickname);
     }
